アクセサ作成ロジック、変換ツールサービスインタフェース、および関連クラスを追加
