+++ src/main/java/kmg/tool/application/logic/io/AccessorCreationLogic.java
@@ -1,0 +1,12 @@
+package kmg.tool.application.logic.io;
+
+/**
+ * アクセサ作成ロジックインタフェース<br>
+ *
+ * @author KenichiroArai
+ */
+public interface AccessorCreationLogic {
+
+    // 処理なし
+
+}
