+++ src/main/java/kmg/tool/application/logic/io/impl/AccessorCreationLogicImpl.java
@@ -1,0 +1,14 @@
+package kmg.tool.application.logic.io.impl;
+
+import kmg.tool.application.logic.io.AccessorCreationLogic;
+
+/**
+ * アクセサ作成ロジック<br>
+ *
+ * @author KenichiroArai
+ */
+public class AccessorCreationLogicImpl implements AccessorCreationLogic {
+
+    // 処理なし
+
+}
