+++ src/main/java/kmg/tool/application/service/io/AccessorCreationService.java
@@ -1,0 +1,14 @@
+package kmg.tool.application.service.io;
+
+import kmg.tool.domain.service.io.Two2OneService;
+
+/**
+ * アクセサ作成サービスインタフェース<br>
+ *
+ * @author KenichiroArai
+ */
+public interface AccessorCreationService extends Two2OneService {
+
+    // 処理なし
+
+}
