+++ src/main/java/kmg/tool/application/service/io/SimpleOne2OneService.java
@@ -1,0 +1,12 @@
+package kmg.tool.application.service.io;
+
+import kmg.tool.domain.service.io.One2OneService;
+
+/**
+ * シンプル1入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
+ */
+public interface SimpleOne2OneService extends One2OneService {
+
+    // 処理なし
+
+}
