+++ src/main/java/kmg/tool/application/service/io/SimpleTwo2OneService.java
@@ -1,0 +1,12 @@
+package kmg.tool.application.service.io;
+
+import kmg.tool.domain.service.io.Two2OneService;
+
+/**
+ * シンプル2入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
+ */
+public interface SimpleTwo2OneService extends Two2OneService {
+
+    // 処理なし
+
+}
