+++ src/main/java/kmg/tool/application/service/io/impl/SimpleTwo2OneServiceImpl.java
@@ -1,0 +1,178 @@
+package kmg.tool.application.service.io.impl;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Service;
+
+import kmg.core.infrastructure.types.KmgDelimiterTypes;
+import kmg.tool.application.service.io.SimpleTwo2OneService;
+import kmg.tool.domain.types.KmgToolGenMessageTypes;
+import kmg.tool.infrastructure.exception.KmgToolException;
+
+/**
+ * シンプル2入力ファイルから1出力ファイルへの変換ツールサービス<br>
+ */
+@Service
+public class SimpleTwo2OneServiceImpl implements SimpleTwo2OneService {
+
+    /** 入力ファイルパス */
+    private Path inputPath;
+
+    /** テンプレートファイルパス */
+    private Path templatePath;
+
+    /** 出力ファイルパス */
+    private Path outputPath;
+
+    /**
+     * 入力ファイルパスを返す<br>
+     *
+     * @author KenichiroArai
+     *
+     * @sine 1.0.0
+     *
+     * @version 1.0.0
+     *
+     * @return 入力ファイルパス
+     */
+    @Override
+    public Path getInputPath() {
+
+        final Path result = this.inputPath;
+        return result;
+
+    }
+
+    /**
+     * 出力ファイルパスを返す<br>
+     *
+     * @author KenichiroArai
+     *
+     * @sine 1.0.0
+     *
+     * @version 1.0.0
+     *
+     * @return 出力ファイルパス
+     */
+    @Override
+    public Path getOutputPath() {
+
+        final Path result = this.outputPath;
+        return result;
+
+    }
+
+    /**
+     * テンプレートファイルパスを返す<br>
+     *
+     * @author KenichiroArai
+     *
+     * @sine 1.0.0
+     *
+     * @version 1.0.0
+     *
+     * @return テンプレートファイルパス
+     */
+    @Override
+    public Path getTemplatePath() {
+
+        final Path result = this.templatePath;
+        return result;
+
+    }
+
+    /**
+     * 初期化する
+     *
+     * @return true：成功、false：失敗
+     *
+     * @param inputPath
+     *                     入力ファイルパス
+     * @param templatePath
+     *                     テンプレートファイルパス
+     * @param outputPath
+     *                     出力ファイルパス
+     */
+    @SuppressWarnings("hiding")
+    @Override
+    public boolean initialize(final Path inputPath, final Path templatePath, final Path outputPath) {
+
+        final boolean result = true;
+
+        this.inputPath = inputPath;
+        this.templatePath = templatePath;
+        this.outputPath = outputPath;
+
+        return result;
+
+    }
+
+    /**
+     * 処理する
+     *
+     * @return true：成功、false：失敗
+     *
+     * @throws KmgToolException
+     *                          KMGツール例外
+     */
+    @Override
+    public boolean process() throws KmgToolException {
+
+        boolean result = false;
+
+        /* テンプレートの取得 */
+        String template = null;
+
+        try {
+
+            template = Files.readAllLines(this.templatePath).stream()
+                .collect(Collectors.joining(KmgDelimiterTypes.LINE_SEPARATOR.get()));
+
+        } catch (final IOException e) {
+
+            // 例外をスローする
+            final KmgToolGenMessageTypes msgType     = KmgToolGenMessageTypes.KMGTOOL_GEN31002;
+            final Object[]               messageArgs = {
+                this.templatePath.toString()
+            };
+            throw new KmgToolException(msgType, messageArgs, e);
+
+        }
+
+        /* 入力から出力の処理 */
+        try (final BufferedReader brInput = Files.newBufferedReader(this.inputPath);
+            final BufferedWriter bw = Files.newBufferedWriter(this.outputPath);) {
+
+            final StringBuilder output = new StringBuilder();
+
+            String line = null;
+
+            while ((line = brInput.readLine()) != null) {
+
+                final String wk = template.replace("{ name }", line);
+                output.append(wk);
+
+            }
+
+            bw.write(output.toString());
+
+        } catch (final IOException e) {
+
+            // 例外をスローする
+            final KmgToolGenMessageTypes msgType     = KmgToolGenMessageTypes.KMGTOOL_GEN31001;
+            final Object[]               messageArgs = {};
+            throw new KmgToolException(msgType, messageArgs, e);
+
+        }
+
+        result = true;
+        return result;
+
+    }
+
+}
