+++ src/main/java/kmg/tool/domain/service/io/DynamicTemplateConversionService.java
@@ -1,0 +1,10 @@
+package kmg.tool.domain.service.io;
+
+/**
+ * テンプレートの動的変換サービスインタフェース
+ */
+public interface DynamicTemplateConversionService extends Two2OneService {
+
+    // 処理なし
+
+}
