--- src/main/java/kmg/tool/io/application/logic/AccessorCreationLogic.java
@@ -1,12 +1,0 @@
-package kmg.tool.io.application.logic;
-
-/**
- * アクセサ作成ロジックインタフェース<br>
- *
- * @author KenichiroArai
- */
-public interface AccessorCreationLogic {
-
-    // 処理なし
-
-}
