--- src/main/java/kmg/tool/io/application/logic/impl/AccessorCreationLogicImpl.java
@@ -1,14 +1,0 @@
-package kmg.tool.io.application.logic.impl;
-
-import kmg.tool.io.application.logic.AccessorCreationLogic;
-
-/**
- * アクセサ作成ロジック<br>
- *
- * @author KenichiroArai
- */
-public class AccessorCreationLogicImpl implements AccessorCreationLogic {
-
-    // 処理なし
-
-}
