--- src/main/java/kmg/tool/io/application/service/AccessorCreationService.java
@@ -1,14 +1,0 @@
-package kmg.tool.io.application.service;
-
-import kmg.tool.io.domain.service.Two2OneService;
-
-/**
- * アクセサ作成サービスインタフェース<br>
- *
- * @author KenichiroArai
- */
-public interface AccessorCreationService extends Two2OneService {
-
-    // 処理なし
-
-}
