--- src/main/java/kmg/tool/io/application/service/SimpleOne2OneService.java
@@ -1,12 +1,0 @@
-package kmg.tool.io.application.service;
-
-import kmg.tool.io.domain.service.One2OneService;
-
-/**
- * シンプル1入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
- */
-public interface SimpleOne2OneService extends One2OneService {
-
-    // 処理なし
-
-}
