--- src/main/java/kmg/tool/io/application/service/SimpleTwo2OneService.java
@@ -1,12 +1,0 @@
-package kmg.tool.io.application.service;
-
-import kmg.tool.io.domain.service.Two2OneService;
-
-/**
- * シンプル2入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
- */
-public interface SimpleTwo2OneService extends Two2OneService {
-
-    // 処理なし
-
-}
