--- src/main/java/kmg/tool/io/domain/service/DynamicTemplateConversionService.java
@@ -1,10 +1,0 @@
-package kmg.tool.io.domain.service;
-
-/**
- * テンプレートの動的変換サービスインタフェース
- */
-public interface DynamicTemplateConversionService extends Two2OneService {
-
-    // 処理なし
-
-}
