--- src/main/java/kmg/tool/io/domain/service/Two2OneService.java
@@ -1,67 +1,0 @@
-package kmg.tool.io.domain.service;
-
-import java.nio.file.Path;
-
-import kmg.tool.infrastructure.exception.KmgToolException;
-
-/**
- * シンプル2入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
- */
-public interface Two2OneService extends IoService {
-
-    /**
-     * 入力ファイルパスを返す<br>
-     *
-     * @author KenichiroArai
-     *
-     * @sine 1.0.0
-     *
-     * @version 1.0.0
-     *
-     * @return 入力ファイルパス
-     */
-    Path getInputPath();
-
-    /**
-     * 出力ファイルパスを返す<br>
-     *
-     * @author KenichiroArai
-     *
-     * @sine 1.0.0
-     *
-     * @version 1.0.0
-     *
-     * @return 出力ファイルパス
-     */
-    Path getOutputPath();
-
-    /**
-     * テンプレートファイルパスを返す<br>
-     *
-     * @author KenichiroArai
-     *
-     * @sine 1.0.0
-     *
-     * @version 1.0.0
-     *
-     * @return テンプレートファイルパス
-     */
-    Path getTemplatePath();
-
-    /**
-     * 初期化する
-     *
-     * @return true：成功、false：失敗
-     *
-     * @param inputPath
-     *                     入力ファイルパス
-     * @param templatePath
-     *                     テンプレートファイルパス
-     * @param outputPath
-     *                     出力ファイルパス
-     *
-     * @throws KmgToolException
-     *                          KMGツール例外
-     */
-    boolean initialize(final Path inputPath, final Path templatePath, final Path outputPath) throws KmgToolException;
-}
